Strings now moved to a new parts_etc schema.
After some schema sketching, and some thought, I think the 10 or so tables that strings take up don't require their own schema.  It makes more sense to me to group them with all the parts and accessories needed for playing an instrument.  Examples of similar things that come to mind are reeds and pads for woodwinds, which are also expected to be worn out and replaced over the course of standard use of the instrument.