Table naming change. Strings is a special case.
I started naming tables in plural, because string is a reserved word in SQL, but strings is not.  Wanting to keep table names consistent, I used plurals.  But, this isn't really feeling right, so I'm renaming all the tables except strings to singular.  I might rename strings to instrument_string, but right now I'm going to leave it as a special case.